Move getPath from expect/utils to jest-util
