+++ packages/jest-each/src/bind.js
@@ -10,6 +10,7 @@
 import util from 'util';
 import chalk from 'chalk';
 import pretty from 'pretty-format';
+import {getPath} from 'jest-util';
 
 type Table = Array<Array<any>>;
 type PrettyArgs = {
@@ -152,53 +153,3 @@ const applyObjectParams = (obj: any, test: Function) => {
 
 const pluralize = (word: string, count: number) =>
   word + (count === 1 ? '' : 's');
-
-const hasOwnProperty = (object: Object, value: string) =>
-  Object.prototype.hasOwnProperty.call(object, value) ||
-  Object.prototype.hasOwnProperty.call(object.constructor.prototype, value);
-
-const getPath = (object: Object, propertyPath: string | Array<string>) => {
-  if (!Array.isArray(propertyPath)) {
-    propertyPath = propertyPath.split('.');
-  }
-
-  if (propertyPath.length) {
-    const lastProp = propertyPath.length === 1;
-    const prop = propertyPath[0];
-    const newObject = object[prop];
-
-    if (!lastProp && (newObject === null || newObject === undefined)) {
-      // This is not the last prop in the chain. If we keep recursing it will
-      // hit a `can't access property X of undefined | null`. At this point we
-      // know that the chain has broken and we can return right away.
-      return {
-        hasEndProp: false,
-        lastTraversedObject: object,
-        traversedPath: [],
-      };
-    }
-
-    const result = getPath(newObject, propertyPath.slice(1));
-
-    if (result.lastTraversedObject === null) {
-      result.lastTraversedObject = object;
-    }
-
-    result.traversedPath.unshift(prop);
-
-    if (lastProp) {
-      result.hasEndProp = hasOwnProperty(object, prop);
-      if (!result.hasEndProp) {
-        result.traversedPath.shift();
-      }
-    }
-
-    return result;
-  }
-
-  return {
-    lastTraversedObject: null,
-    traversedPath: [],
-    value: object,
-  };
-};
