+++ packages/jest-util/src/get_path.js
@@ -1,0 +1,61 @@
+type GetPath = {
+  hasEndProp?: boolean,
+  lastTraversedObject: ?Object,
+  traversedPath: Array<string>,
+  value?: any,
+};
+
+const hasOwnProperty = (object: Object, value: string) =>
+  Object.prototype.hasOwnProperty.call(object, value) ||
+  Object.prototype.hasOwnProperty.call(object.constructor.prototype, value);
+
+const getPath = (
+  object: Object,
+  propertyPath: string | Array<string>,
+): GetPath => {
+  if (!Array.isArray(propertyPath)) {
+    propertyPath = propertyPath.split('.');
+  }
+
+  if (propertyPath.length) {
+    const lastProp = propertyPath.length === 1;
+    const prop = propertyPath[0];
+    const newObject = object[prop];
+
+    if (!lastProp && (newObject === null || newObject === undefined)) {
+      // This is not the last prop in the chain. If we keep recursing it will
+      // hit a `can't access property X of undefined | null`. At this point we
+      // know that the chain has broken and we can return right away.
+      return {
+        hasEndProp: false,
+        lastTraversedObject: object,
+        traversedPath: [],
+      };
+    }
+
+    const result = getPath(newObject, propertyPath.slice(1));
+
+    if (result.lastTraversedObject === null) {
+      result.lastTraversedObject = object;
+    }
+
+    result.traversedPath.unshift(prop);
+
+    if (lastProp) {
+      result.hasEndProp = hasOwnProperty(object, prop);
+      if (!result.hasEndProp) {
+        result.traversedPath.shift();
+      }
+    }
+
+    return result;
+  }
+
+  return {
+    lastTraversedObject: null,
+    traversedPath: [],
+    value: object,
+  };
+};
+
+export default getPath;
